Setup.py added
Setup.py added, code.py renamed, Vowel_Test moved under test file. Commit to test the setup.
