+++ code/Vowel_Check.py
@@ -1,0 +1,8 @@
+def vowel(c):
+    x=['a','A','e','E','i','I','o','O','u','U']
+    if(c in x):
+        print(c," is vowel")
+        return 1
+    else:
+        print(c," is not vowel")
+        return 0
